Simplify endpoint extraction in ConnectionSettings
