+++ src/Ev.ServiceBus.Abstractions/Configuration/ClientOptions.cs
@@ -1,5 +1,7 @@
 ﻿// ReSharper disable once CheckNamespace
 
+using Ev.ServiceBus.Abstractions.Configuration;
+
 namespace Ev.ServiceBus.Abstractions;
 
 public abstract class ClientOptions : IClientOptions
