+++ src/Ev.ServiceBus.Abstractions/Configuration/Extensions/ClientOptionsExtensions.cs
@@ -1,4 +1,5 @@
 ﻿using Azure.Messaging.ServiceBus;
+using Ev.ServiceBus.Abstractions.Configuration;
 
 // ReSharper disable once CheckNamespace
 namespace Ev.ServiceBus.Abstractions;
