+++ src/Ev.ServiceBus.Abstractions/Configuration/IClientOptions.cs
@@ -1,4 +1,7 @@
 ﻿// ReSharper disable once CheckNamespace
+
+using Ev.ServiceBus.Abstractions.Configuration;
+
 namespace Ev.ServiceBus.Abstractions;
 
 public interface IClientOptions
