+++ src/Ev.ServiceBus.Abstractions/Configuration/ServiceBusSettings.cs
@@ -1,4 +1,5 @@
 ﻿using Azure.Messaging.ServiceBus;
+using Ev.ServiceBus.Abstractions.Configuration;
 
 namespace Ev.ServiceBus.Abstractions;
