+++ src/Ev.ServiceBus.HealthChecks/ConnectionSettingsComparer.cs
@@ -1,6 +1,5 @@
-﻿using System;
-using System.Collections.Generic;
-using Ev.ServiceBus.Abstractions;
+﻿using System.Collections.Generic;
+using Ev.ServiceBus.Abstractions.Configuration;
 
 namespace Ev.ServiceBus.HealthChecks;
