+++ src/Ev.ServiceBus/Management/Factories/ClientFactory.cs
@@ -1,5 +1,5 @@
 using Azure.Messaging.ServiceBus;
-using Ev.ServiceBus.Abstractions;
+using Ev.ServiceBus.Abstractions.Configuration;
 
 namespace Ev.ServiceBus;
