+++ src/Ev.ServiceBus/Management/Wrappers/ComposedReceiverOptions.cs
@@ -2,6 +2,7 @@
 using System.Linq;
 using Azure.Messaging.ServiceBus;
 using Ev.ServiceBus.Abstractions;
+using Ev.ServiceBus.Abstractions.Configuration;
 
 namespace Ev.ServiceBus;
