+++ tests/Ev.ServiceBus.TestHelpers/FailingClientFactory.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using Azure.Messaging.ServiceBus;
-using Ev.ServiceBus.Abstractions;
+using Ev.ServiceBus.Abstractions.Configuration;
 
 namespace Ev.ServiceBus.UnitTests.Helpers;
