+++ tests/Ev.ServiceBus.TestHelpers/FakeClientFactory.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 using Azure.Messaging.ServiceBus;
-using Ev.ServiceBus.Abstractions;
+using Ev.ServiceBus.Abstractions.Configuration;
 
 namespace Ev.ServiceBus.TestHelpers;
